+++ contracts/finality/src/state/mod.rs
@@ -1,5 +1,6 @@
 pub mod config;
 pub mod finality;
+pub mod pruning;
 pub mod public_randomness;
 
 /// `Bytes` is a vector of bytes